Move RedwoodTester to redwood-compose-testing
