+++ redwood-compose-testing/gradle.properties
@@ -1,0 +1,2 @@
+POM_ARTIFACT_ID=redwood-compose-testing
+POM_NAME=Redwood Compose Testing
