+++ redwood-compose-testing/src/commonMain/kotlin/app/cash/redwood/compose/testing/Semaphores.kt
@@ -14,16 +14,14 @@
  * limitations under the License.
  */
 
-package com.example.redwood.emojisearch.presenter
+package app.cash.redwood.compose.testing
 
-import app.cash.redwood.RedwoodCodegenApi
 import kotlinx.coroutines.sync.Semaphore
 
-@RedwoodCodegenApi
-class ChangeTracker {
-  internal val changes = Semaphore(Int.MAX_VALUE, Int.MAX_VALUE)
-
-  public fun widgetChanged() {
-    changes.release()
+internal fun Semaphore.acquireAll(): Int {
+  var result = 0
+  while (tryAcquire()) {
+    result++
   }
+  return result
 }
