+++ settings.gradle
@@ -2,6 +2,7 @@ rootProject.name = 'redwood'
 
 include ':redwood-cli'
 include ':redwood-compose'
+include ':redwood-compose-testing'
 include ':redwood-flexbox'
 include ':redwood-gradle-plugin'
 include ':redwood-layout-api'
